CUMULUS-2309 refactor granule delete library function to separate file
